planner refactor cont'd
